+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added support for query `GetAncestors`
 - Added support for query `JoinNetwork`
 - Added support for query `LeaveNetwork`
+- Added support for query `GetInstanceInfo`
 - Added support for query `GetInstances`
 - Added support for query `GetAccountList`
 - Added support for query `GetBannedPeers`